Moved two classes out of Comm.java and into their own files.
